pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "docling>=2.80.0",
    "filelock>=3.25.2",
    "google-genai>=1.67.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
    "tavily-python>=0.7.23",
    "tenacity>=9.1.4",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "requests-mock>=1.12.1",
    "ruff>=0.15.6",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
